% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_waittimes.R
\name{get_wait_times}
\alias{get_wait_times}
\title{Get Magic Kingdom Wait Times}
\usage{
get_wait_times()
}
\value{
tibble with Magic Kingdom wait times
}
\description{
Get Magic Kingdom Wait Times
}
\examples{
\dontrun{
get_wait_times()
}
}
